Add Python 3.10 compatibility This update adds Python 3.10 to epispot's testing suite and makes it part of the default environment. After this change, Python 3.10 should be used by all GitHub runners and will also be our primary focus for testing. See the additional notes for update-specific changes.
Changelist:
- .github/workflows
    - Use Python 3.10 as the default runner environment
    - build.yml
        - Add Python 3.10 to testing matrix
- requirements/
    - Use absolute latest version of matplotlib to avoid deprecation
      warnings on Python 3.10 (more on this in the additional notes)

Additional Notes:

Deprecation Notice:
Despite the changes announced in this update, we are *not* deprecating
any of the previous
Python versions, including Python 3.7.

Dependency Conflicts:
Due to the goal of making Python 3.10 the default interpreter for the
epispot package, we've had to use a release candidate version of
matplotlib (3.5.0rc1) instead of a stable one.
This is done to avoid a deprecation warning which occurs because the
latest stable release still uses distutils, which is deprecated in
Python 3.10.